+++ src/app.ts
@@ -6,10 +6,8 @@ import AppError from './middleware/errors/AppError.js';
 import { sendSuccess } from './utils/response.js';
 import { requireAuth } from './middleware/authMiddleware.js';
 import authRoutes from './routes/authRoutes.js';
-import youtubeRoutes from './routes/youtubeRecommendationRoute.js';
-import youtubeSearchRouter from './routes/youtubeSearchRoute.js';
+import youtubeRoutes from './routes/youtubeRoute.js';
 import userRoutes from './routes/userRoutes.js';
-import friendRoutes from './routes/friendRoutes.js';
 import swaggerUi from 'swagger-ui-express';
 import { specs } from './swagger.js';
 import { createServer } from 'http';
@@ -46,25 +44,20 @@ const corsOptions = {
 
     // 프로덕션 환경에서는 허용된 도메인만
     const allowedOrigins = [
-      //수정1 - undefined 값 필터링
+      //수정1
       address,
-      'http://54.180.254.48:3000', // HTTP로 수정
       'https://54.180.254.48:3000',
       //'https://your-frontend-domain.com', // 실제 프론트엔드 도메인으로 변경
       //'https://onairmate.vercel.app', // 예시 도메인
       'http://localhost:3000', // 로컬 개발용
       'http://localhost:3001', // 로컬 개발용
-    ].filter(Boolean); // undefined나 null 값 제거
-
+    ];
     console.log('배포 주소', address);
     console.log('연결 origin:', origin);
 
-    // origin이 없는 경우(same-origin 요청) 또는 허용된 origin인 경우
-    if (!origin || allowedOrigins.includes(origin)) {
+    if (!origin || allowedOrigins.indexOf(origin) !== -1) {
       callback(null, true);
     } else {
-      console.log('CORS 거부됨:', origin);
-      console.log('허용된 origins:', allowedOrigins);
       callback(new Error('Not allowed by CORS'));
     }
   },
@@ -141,12 +134,9 @@ app.get('/', (req: Request, res: Response) => {
 // API 라우트들을 여기에 추가
 app.use('/api/auth', authRoutes);
 app.use('/api/users', userRoutes);
-app.use('/api/friends', friendRoutes);
 app.use('/api/rooms', roomRoutes);
 app.use('/api/chat/direct', chatDirectRoutes);
-app.use('/api/youtube', youtubeRoutes); // youtubeRecommendationRoute.js 와 youtubeSearchRoute.js 를 병합하여 사용해야 합니다.
-app.use('/api/youtube', youtubeRoutes);
-app.use('/api/youtube', youtubeSearchRouter);
+app.use('/api/youtube', youtubeRoutes); // youtubeRecommendationRoute와 youtubeSearchRoute 병합
 
 // 404 에러 핸들링
 app.use((req: Request, res: Response, next: NextFunction) => {
